New umbral.random_oracles module to gather all hash and KDF functions
